modularize database connection export and createdb
